refactor: migrate pages to src dir
