+++ src/pages/showcase/charts/GithubPRExample.tsx
@@ -19,8 +19,8 @@ import { BrothButton } from '@/components/button/BrothButton';
 import { Stack } from '@mui/system';
 import { Octokit } from 'octokit';
 import { format } from 'date-fns';
-import { formatIsoDate } from '../../../helpers/dateHelpers';
-import { BrothTypography } from '../../../components/typography/BrothTypography';
+import { formatIsoDate } from '../../../../helpers/dateHelpers';
+import { BrothTypography } from '../../../../components/typography/BrothTypography';
 
 type ChartData = {
   date: string;
