+++ src/pages/showcase/dashboards/FinancialDashboardExample.tsx
