+++ src/pages/showcase/heroes/HeroSignup.tsx
