+++ src/pages/showcase/heroes/HeroVisual.tsx
