+++ stories/examples/Forms.stories.tsx
@@ -1,8 +1,8 @@
-import AddressFormExample from 'pages/showcase/forms/AddressFormExample';
-import MultiStepFormExample from 'pages/showcase/forms/MultiStepFormExample';
-import ValidatedFormExample from 'pages/showcase/forms/ValidatedFormExample';
-import ImageUploadFormExample from 'pages/showcase/forms/ImageUploadFormExample';
-import MultipleImageFormExample from 'pages/showcase/forms/MultipleImageFormExample';
+import AddressFormExample from 'src/pages/showcase/forms/AddressFormExample';
+import MultiStepFormExample from 'src/pages/showcase/forms/MultiStepFormExample';
+import ValidatedFormExample from 'src/pages/showcase/forms/ValidatedFormExample';
+import ImageUploadFormExample from 'src/pages/showcase/forms/ImageUploadFormExample';
+import MultipleImageFormExample from 'src/pages/showcase/forms/MultipleImageFormExample';
 import { showcaseParameters } from './showcaseParameters';
 
 export default {
